fix(recovery): isolate resume event prefix
