+++ pkg/services/factory_runtime/internal/runtime_build.go
@@ -221,8 +221,8 @@ func (service runtimeWorkersServiceWithProgress) ResolveTemplateFields(
 
 type runtimeOpeningWithFlush struct {
 	recordings.RuntimeOpening
-	flushInterval time.Duration
-	replayEvents  []factorydefinitions.FactoryEvent
+	flushInterval         time.Duration
+	resumeCanonicalEvents []factorydefinitions.FactoryEvent
 }
 
 func (opening runtimeOpeningWithFlush) OpenRuntime(
@@ -230,7 +230,7 @@ func (opening runtimeOpeningWithFlush) OpenRuntime(
 	request recordings.RuntimeScopeRequest,
 ) (recordings.RuntimeScopeResult, error) {
 	request.FlushInterval = opening.flushInterval
-	request.ReplayEvents = cloneFactoryEvents(opening.replayEvents)
+	request.ReplayEvents = cloneFactoryEvents(opening.resumeCanonicalEvents)
 	return opening.RuntimeOpening.OpenRuntime(ctx, request)
 }
 
@@ -444,9 +444,9 @@ func buildBundle(
 		spec.PetriMutationRecorder,
 		worldStateProjector,
 		runtimeOpeningWithFlush{
-			RuntimeOpening: recordingsRuntime,
-			flushInterval:  recordFlushInterval,
-			replayEvents:   cloneFactoryEvents(spec.ReplayEvents),
+			RuntimeOpening:        recordingsRuntime,
+			flushInterval:         recordFlushInterval,
+			resumeCanonicalEvents: cloneFactoryEvents(spec.ResumeCanonicalEvents),
 		},
 		workerServiceWithProgress,
 		workerSessionsFactory,
